+++ src/event/economy/market_buy.rs
@@ -6,17 +6,17 @@ pub struct MarketBuy {
     pub timestamp: String,
 
     #[serde(rename = "MarketID")]
-    pub market_id: i64,
+    pub market_id: u64,
 
     #[serde(rename = "Type")]
     pub r#type: String,
 
     #[serde(rename = "Count")]
-    pub count: i64,
+    pub count: u32,
 
     #[serde(rename = "BuyPrice")]
-    pub buy_price: i64,
+    pub buy_price: u32,
 
     #[serde(rename = "TotalCost")]
-    pub total_cost: i64,
+    pub total_cost: u32,
 }
