+++ src/event/economy/shipyard_buy.rs
@@ -12,14 +12,14 @@ pub struct ShipyardBuy {
     pub ship_type_localised: Option<String>,
 
     #[serde(rename = "ShipPrice")]
-    pub ship_price: i64,
+    pub ship_price: u32,
 
     #[serde(rename = "StoreOldShip")]
     pub store_old_ship: String,
 
     #[serde(rename = "StoreShipID")]
-    pub store_ship_id: i64,
+    pub store_ship_id: u64,
 
     #[serde(rename = "MarketID")]
-    pub market_id: i64,
+    pub market_id: u64,
 }
