+++ src/event/inventory/stored_modules.rs
@@ -10,22 +10,22 @@ pub struct StoredModule {
     pub name_localised: String,
 
     #[serde(rename = "StorageSlot")]
-    pub storage_slot: i64,
+    pub storage_slot: u32,
 
     #[serde(rename = "StarSystem")]
     pub star_system: Option<String>,
 
     #[serde(rename = "MarketID")]
-    pub market_id: Option<i64>,
+    pub market_id: Option<u32>,
 
     #[serde(rename = "TransferCost")]
-    pub transfer_cost: Option<i64>,
+    pub transfer_cost: Option<u32>,
 
     #[serde(rename = "TransferTime")]
-    pub transfer_time: Option<i64>,
+    pub transfer_time: Option<u32>,
 
     #[serde(rename = "BuyPrice")]
-    pub buy_price: i64,
+    pub buy_price: u32,
 
     #[serde(rename = "Hot")]
     pub hot: bool,
@@ -34,7 +34,7 @@ pub struct StoredModule {
     pub engineer_modifications: Option<String>,
 
     #[serde(rename = "Level")]
-    pub level: Option<i64>,
+    pub level: Option<u32>,
 
     #[serde(rename = "Quality")]
     pub quality: Option<f64>,
@@ -46,7 +46,7 @@ pub struct StoredModules {
     pub timestamp: String,
 
     #[serde(rename = "MarketID")]
-    pub market_id: i64,
+    pub market_id: u64,
 
     #[serde(rename = "StationName")]
     pub station_name: String,
