+++ src/event/navigation/book_taxi.rs
@@ -5,7 +5,7 @@ pub struct BookTaxi {
     pub timestamp: String,
 
     #[serde(rename = "Cost")]
-    pub cost: i64,
+    pub cost: u32,
 
     #[serde(rename = "DestinationSystem")]
     pub destination_system: String,
